+++ src/main/resources/release_notes.md
@@ -2,6 +2,7 @@
 
 * Calculation of the tsi sub by approximating the tsi formula (#235)
 * Restoring user's table sorting settings (#2119)
+* Display coach conversion costs (#2204)
 
 ## [Detailed Changelog](https://github.com/ho-dev/HattrickOrganizer/issues?q=milestone%3A9.0)
 
@@ -14,6 +15,7 @@
 
 * Calculation of the tsi sub by approximating the tsi formula (#235)
 * Fix display of best position rating (#2166)
+* Display coach conversion costs (#2204)
 
 ### Team Analyzer
 * Fix illegal argument exception in team rating panel (#2155)